feat: add auth framework

+++ samples/SimpleExample/Application/Commands/DeleteWorkspaceCommand.cs
@@ -2,7 +2,10 @@ namespace Api.Application.Commands;
 
 using System.Threading;
 using System.Threading.Tasks;
-using Api.Data;
+using Data;
+using Authorization;
+using Authorization.Abstractions;
+using Keycloak.AuthServices.Authorization.Handlers;
 using MediatR;
 
 public record DeleteWorkspaceCommand(Guid Id) : IRequest;
@@ -10,13 +13,25 @@ public record DeleteWorkspaceCommand(Guid Id) : IRequest;
 public class DeleteWorkspaceCommandHandler : IRequestHandler<DeleteWorkspaceCommand>
 {
     private readonly IApplicationDbContext db;
+    private readonly IIdentityService identityService;
 
-    public DeleteWorkspaceCommandHandler(IApplicationDbContext db) => this.db = db;
+    public DeleteWorkspaceCommandHandler(IApplicationDbContext db, IIdentityService identityService)
+    {
+        this.db = db;
+        this.identityService = identityService;
+    }
 
     public async Task<Unit> Handle(
         DeleteWorkspaceCommand request,
         CancellationToken cancellationToken)
     {
+        var authorized = await this.identityService.AuthorizeAsync(
+            ProtectedResourcePolicy.From("workspaces", request.Id.ToString(), "read"));
+
+        if (!authorized)
+        {
+            throw new ForbiddenAccessException();
+        }
         var workspace = new Workspace() { Id = request.Id };
         this.db.Workspaces.Remove(workspace);
         await this.db.SaveChangesAsync(cancellationToken);
